initial basic source code drop
This adds the initial version of sunxi-fw, a tool to decode firmware
images for Allwinner CPU based systems.
The code so far gives a basic overview of firmware images generated by
mainline U-Boot, beginning with an SPL part, then followed by either a
U-Boot legacy image or a U-Boot FIT image.
It allows the image to be preceeded by a partition table (or MBR).
The extract feature is limited to the MBR and the SPL so far.