add simple TOC0 support
TOC0 is an Allwinner defined image format that is expected by the
BootROM when the chip uses "secure boot", as determined by an eFuse.
In that mode the BootROM will not accept the normal "eGON" image header,
but requires the SPL payload to be wrapped in said TOC0 format, which
additionally carries keys and signatures, to allow verifying the
integrity of a boot image, forming the root of the chain of trust.

Dump very basic information about the TOC0 content, mostly the payload
size, to be able to find other firmware parts in an image.