output functions: return number of read/skipped sectors
Currently the code for each image type works out where the potential
next image starts, and skip forward to that point. That works well, but
misses the information about where in the image we are currently at.

Let each output_*_info() function return the number of sectors skip, so
that the high level code has an idea about the current location in an
image file, and can (later) give this information to the user.