add GPT information
The classic way of partition a drive uses the ancient MS-DOS defined
boot sector and partition table format (MBR), which has nasty limits and
lacks flexibility.

Add support for the more modern GPT partition table format, which is
defined in the UEFI specification and is used on most storage devices
outside of the embedded world.