output generic boot0 information
"boot0" is the name used by the Allwinner BSP SPL replacement, which is
proprietary closed source code that initialises the system and the DRAM
and loads further firmware components. This seems to be custom code, so
there is not much to decode here, but we can at least print the size of
the image and verify the checksum, and skip beyond this stage to find
potentially further firmware components.