support SPL sizes bigger than 32K
For a long time the Allwinner BootROM had a built-in 32K limit, so this
assumption was also baked into some of the code. The BootROM on newer SoCs
can load larger image sizes, and we rely on this for the H616 build of
mainline U-Boot.
Use the same heuristic that mainline U-Boot uses: any SPL size smaller
than 32KB will be padded to 32K, but the SPL payload will follow any
larger SPL directly.
This allows parsing image files for newer SoCs.