boot0: dump embedded DRAM parameter information
Allwinner's proprietary boot0 code (which does a similar job as the
mainline U-Boot SPL) contains board specific DRAM parameters. Apart from
generic information like the DRAM type and clock speed, it also contains
very specific timing parameters which depend on the board layout.

The structure of this information varies with boot0 versions, but
roughly follows SoC generations. We don't know which version a certain
boot0 image uses, but we can dump the information for two popular
formats: one older, used on the H3 and A64 generation, and one newer, as
seen on boot0's for many H616 boards.
For the latter case, the parameters can be directly copied into the
respective Kconfig variables in a mainline U-Boot defconfig file.