drop assumption about fixed SPL size
For a long time the Allwinner BootROM had a built-in 32K limit, so this
assumption was also baked into some of the code. Since the BootROM on
newer SoCs is no longer restricted in the image size it loads, drop the
assumption about this 32K limit which was embedded in the code.