add U-Boot legacy image support
The U-Boot "legacy image" is a simple 64-byte header prepended to a
payload, which contains information about the image type, compression
type, size and load addresses. It was used by U-Boot for a long time,
but has since then been superseeded by the more flexible and capable
U-Boot FIT image format.

The code decodes the header information, and allows to extract the image
or the actual image payload.