dump SPL DT name
The mainline U-Boot SPL contains a versioned header, that can carry
information, either for the SPL itself to use, or to transport from the
SPL to U-Boot proper (like the detected DRAM size).
One information in the header is the DT name that shall be selected from
potentially multiple DTBs stored in the U-Boot FIT image.

Dump the name that is contained in the header, which gives the user a
quick way of determining which board the firmware image has been built
for.