add "dump_dt_names" action
A U-Boot FIT image can contain multiple DTBs, one of which will be
selected by the SPL either by some kind of auto-detection, or by looking
at the DT name embedded in the SPL header.

Add a command to dump the names for each DTB found in the FIT image.