switch to getopt_long()
We don't really need long options, but using getopt_long(), which is a GNU
extension, relaxes the strict requirements that the strict POSIX
compliant musl implementation of getopt() has. This behaviour stops
parsing when it finds the first non-option argument, which is our action
verb. So any parameter afterwards (like -v) is ignored.

The glibc implementation of getopt() and all getopt_long() implementations
shuffle argv and put option arguments first, so work as expected.

This fixes operation on musl based systems, or when compiling a static
binary with musl.