boot0: move dram parameter prints into function
In preparation for adding DRAM.ext support, move the printing of the
DRAM parameters into a reusable function. Change the semantics of OFS so
it is agnostic to the base offset of the parameters. Generate the table
header using a new soc_names array, to enable the simple addition of new
SoCs and improve the header formatting.
Also use the opportunity to print the DRAM clock (the very first
parameter) in decimal.