boot0: add DRAM.ext output support
It was discovered that in the A133's boot0, a flag may be set which
allows for selecting between different DRAM parameters based on the
state of GPIO pins or voltages read by ADC pins on startup. On one
device, this actually appears to be hard-wired to select one of these
configurations.

Start by adding support for printing the DRAM parameters themselves.
The extended header, if present, is located at a fixed address (0x340),
starting with an 8-byte version, followed by the magic "DRAM.ext". The
parameters themselves are stored 36 bytes after. There are additional
pieces of information that could be added later; most notably the method
of determining which channel to use (GPIO, ADC, or both), and which
GPIOs and/or ADC channel are utilized are stored just after the magic.